src/pyproject.toml
[tool.poetry]
name = "horoscope_adviser"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.10.0"
autoconfig = "^5.0.0"
python-decouple = "^3.6"
pydantic = "^2.2.1"
sqlalchemy = {extras = ["mypy"], version = "^2.0.20"}
inflection = "^0.5.1"
alembic = "^1.7.7"
SQLAlchemy-Utils = "^0.38.2"
aiogram = "^3.0"
pydantic-settings = "^2.0.3"
asyncpg = "^0.28.0"
aiohttp = "^3.8.5"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.3.0"
factory-boy = "^3.3.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
